feat: migrate to sepolia feat: cancelbid, allowance for auction
